feat: add tracking of CSS usage and CSS now being removed if no applications are left that use it.
Keep in mind that this PR might have negative affect on parcels mechanism. Usually everything will work as it is now, but there is a corner case when application is mounted and then parcel from the same application is used and application is unmounted. In such a scenario parcel could be left without CSS. Generically speaking since Parcels were not widely adopted within NC, will see if it will break anything. If it breaks something - most-likely we would need to track parcel usages on a page additionally to apps.