fix: case when CSS removed immediately but due to spinner cloned node of application is still alive.
Generic logic of ILC when spinner is enabled is that app is unmounted, but original app node is cloned in DOM and alive while spinner is in progress. That is why if spinner is enabled CSS should not be removed from DOM immediately, instead we need to wait for all transitions and remove CSS only then.